requirements.txt
numpy
tensorflow
pyyaml
tqdm
scikit-learn